+++ packages/terra-framework-docs/src/terra-dev-site/test/tabs/TabPane/LongTextTabPane.test.jsx
@@ -1,0 +1,17 @@
+import React from 'react';
+import classNames from 'classnames/bind';
+import TabPane from 'terra-tabs/lib/TabPane';
+import styles from './common/TabPane.test.module.scss';
+
+const cx = classNames.bind(styles);
+
+const LongTextTabPane = () => (
+  <div role="tablist" className={cx('content-wrapper')}>
+    <TabPane
+      label="Lorem ipsum dolor sit amet, consectetur adipiscing elit. Sed volutpat eros quis erat tincidunt, id scelerisque magna ultricies. Vivamus pellentesque."
+      id="longText"
+    />
+  </div>
+);
+
+export default LongTextTabPane;