+++ packages/terra-theme-provider/tests/wdio/theme-provider-spec.js
@@ -2,23 +2,23 @@ const selector = '#root';
 
 Terra.describeViewports('Theme Provider', ['tiny', 'medium'], () => {
   it('Displays a default theme-provider', () => {
-    browser.url('/raw/tests/terra-theme-provider/theme-provider/default-theme-provider');
+    browser.url('/raw/tests/cerner-terra-framework-docs/theme-provider/default-theme-provider');
     Terra.validates.element('default', { rules: { 'color-contrast': { enabled: false } }, selector });
   });
 
   it('Displays an updated themed component', () => {
-    browser.url('/raw/tests/terra-theme-provider/theme-provider/switch-themes');
+    browser.url('/raw/tests/cerner-terra-framework-docs/theme-provider/switch-themes');
     Terra.validates.element('updated themed component', { selector });
   });
 
   it('Displays an updated themed component - After Click', () => {
-    browser.url('/raw/tests/terra-theme-provider/theme-provider/switch-themes');
+    browser.url('/raw/tests/cerner-terra-framework-docs/theme-provider/switch-themes');
     $('#theme').selectByAttribute('value', 'cerner-mock-theme');
     Terra.validates.element('updated themed component after click', { rules: { 'color-contrast': { enabled: false } }, selector });
   });
 
   it('Displays a themed provider without a themeName provided', () => {
-    browser.url('/raw/tests/terra-theme-provider/theme-provider/theme-provider-no-theme');
+    browser.url('/raw/tests/cerner-terra-framework-docs/theme-provider/theme-provider-no-theme');
     Terra.validates.element('no themeName', { selector });
   });
 });