+++ packages/terra-theme-provider/CHANGELOG.md
@@ -1,6 +1,8 @@
 # Changelog
 
 ## Unreleased
+  * Added
+    * Migrate docs to terra-framework-docs
 
 ## 4.10.0 - (September 20, 2021)